+++ CHANGES.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## 2.1.0
+
+- Feature authenticate via bear token vs username and password
+- Feature added action find_user
+
 ## 2.0.0
 
 - Fixed bug in entry point for send_log.